+++ Editor/Services/VPMService.cs
@@ -1,0 +1,36 @@
+using System.Net.Http;
+using System.Threading.Tasks;
+using KRT.VRCQuestTools.Models;
+using Unity.Plastic.Newtonsoft.Json;
+
+namespace KRT.VRCQuestTools.Services
+{
+    /// <summary>
+    /// VPM operations.
+    /// </summary>
+    internal class VPMService
+    {
+        private static readonly HttpClient Client = new HttpClient();
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="VPMService"/> class.
+        /// </summary>
+        internal VPMService()
+        {
+            Client.Timeout = System.TimeSpan.FromSeconds(10);
+            Client.DefaultRequestHeaders.Add("User-Agent", VRCQuestTools.Name);
+        }
+
+        /// <summary>
+        /// Get VPM repository.
+        /// </summary>
+        /// <param name="url">VPM repository URL.</param>
+        /// <returns>Parsed VPM repository.</returns>
+        internal async Task<VPMRepository> GetVPMRepository(string url)
+        {
+            var result = await Client.GetStringAsync(url);
+            var repo = JsonConvert.DeserializeObject<VPMRepository>(result);
+            return repo;
+        }
+    }
+}